Combine executor_slot_state and function_slot_state.
Instead of having a polymorphic slot_state with two implementations,
slot_state is now a concrete class with a type-erased executor. If a
connection is made without an executor then the slot_state will be
assigned an inline_executor which will simply execute the closure
inline rather than deferring it.