Forward arguments when only a single slot is connected.
This is an optimisation for the case when there is only a single slot
connected to a signal, which is expected to be the common case. When
there is only a single target for the signal's parameters then we can
forward them (as in perfect forwarding) directly to it without making
any copies. This reduces the number of copies required (although
unfortunately not to zero).