Add function to connect an emitter to a signal.
Remove the `emitter::create_signal()` factory function and replace it
with a `connect()` free function which will associate an existing
emitter with an existing signal. This allows the user a little more
flexibility when creating signals, and also means that it's now
possible to publically create an empty signal or an empty slot.