Refactor slot creation and connection.
Creating a slot during a call to `signal::connect()` had some
weaknesses, one being that it made it impossible for a client to
expose a slot for another client to connect to a signal (i.e. public
slots weren't possible).

To address this, the `signal::connect()` functions have been removed
altogether and replaced with a `connect()` free function which
connects a signal to an existing slot. This means that a slot can now
be created with a function but without a connection to an associated
signal. Another consequence is that the slot class is now templated on
the parameter types, which is acceptable as the client must know them
in order to be able to create a connection.

As an example, the following code is now possible:

    class Foo
    {
    public:
      ...
      bb::slot<int> exit;
      ...
    };

    class Bar
    {
    public:
      ...
      bb::signal<int> onExit;
      ...
    };

    int main()
    {
      Foo foo;
      Bar bar;
      bb::connect(bar.onExit, foo.exit);
      return 0;
    }