Allow a function to connect directly to a slot.
Add a new `bb::connect()` overload to connect a function or function
object to a signal without requiring a slot. This is useful when the
client wants to receive the signals for the entire lifetime of the
emitter. Internally this works by creating and storing persistent
slot_states within signal_state, and connecting them as usual.