Stop allowing signal_state to outlive its signal.
Both emitter and signal were maintaining strong ownership of the
signal_state via shared_ptrs. This meant that the signal_state would
live on even after the signal had gone out of scope. Although it would
be impossible to make any more connections to the signal, it was still
possible for the slots to receive the signals directly from the
emitter. In order to fix this, the emitter now only holds a weak_ptr
to the signal_state.